% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{geno_low2}
\alias{geno_low2}
\title{Data from Jiang et al (2021)}
\format{
A matrix containing the genotypes of 121 autotetraploid
individuals at 228,366 snps. The rows index the individuals
and the columns index the genes.
}
\source{
\url{https://github.com/LiboJiang/DoubleReduction}
}
\usage{
geno_low2
}
\description{
Data from Jiang et al (2021)
}
\references{
\itemize{
\item{Jiang, Libo, Xiangyu Ren, and Rongling Wu. 2021. "Computational Characterization of Double Reduction in Autotetraploid Natural Populations." \emph{The Plant Journal} 105 (6): 1703–9. \doi{10.1111/tpj.15126}}
}
}
\keyword{datasets}
