% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi.R
\name{main_multi}
\alias{main_multi}
\title{Fit \code{\link{main_p2}()} across many SNPs.}
\usage{
main_multi(nmat)
}
\arguments{
\item{nmat}{A matrix of counts where the columns index the genotypes and
the rows index the loci. \code{nmat[i, j]} contains the number
of individuals with genotype \code{j-1} at locus \code{i}.}
}
\value{
A matrix with two columns. The first column contains the estimated
allele frequency, the second column contains the estimated double
reduction parameter. The rows index the loci.
}
\description{
Support is provided for parallelization through the future package.
}
\examples{
## Fake data
nmat <- matrix(c(1, 2, 3, 2, 1,
                 3, 2, 1, 2, 3),
               nrow = 2,
               byrow = TRUE)

## Run the analysis in parallel on the local computer with two workers
future::plan(future::multisession, workers = 2)
main_multi(nmat = nmat)

## Shut down parallel workers
future::plan("sequential")

}
\references{
\itemize{
\item{Jiang, Libo, Xiangyu Ren, and Rongling Wu. 2021. "Computational Characterization of Double Reduction in Autotetraploid Natural Populations." \emph{The Plant Journal} 105 (6): 1703–9. \doi{10.1111/tpj.15126}}
}
}
\author{
David Gerard
}
