% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dr_M1.R
\name{main_p2}
\alias{main_p2}
\title{Estimate allele frequency and double reduction in tetraploids using
all genotypes.}
\usage{
main_p2(nm, oiter = 10)
}
\arguments{
\item{nm}{A vector of genotypes. \code{nm[[i]]} is the genotype
for individual \code{i}.}

\item{oiter}{The number of EM iterations to run.}
}
\description{
This uses all genotypes, as opposed to \code{\link{main_p3}()},
which only uses homozygosity versus heterzygosity.
}
\examples{
nm <- sim_p(c(0.5, 0.1), n = 100)
main_p2(nm = nm)

}
\author{
Libo Jiang
}
