% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dr_M2.R
\name{main_p3}
\alias{main_p3}
\title{Estimate allele frequency and double reduction in tetraploids using
just homozygosity versus heterozygosity.}
\usage{
main_p3(nm, oiter = 10)
}
\arguments{
\item{nm}{A vector of genotypes. \code{nm[[i]]} is the genotype
for individual \code{i}. This is coded weird, though. You should
set 0 = nullplex, 1 = heterozygous, 2 = quadriplex.}

\item{oiter}{The number of EM iterations to run.}
}
\value{
A vector of length 2. The first element is the
estimated allele frequency, the second element is the
estimated double reduction parameter.
}
\description{
This uses just homozygosity versus heterozygosity, as opposed to
\code{\link{main_p2}()}, which uses all genotypes.
}
\examples{
nm <- sim_p3(c(0.5, 0.1), n = 100)
main_p3(nm = nm)

}
\author{
Libo Jiang
}
