% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dr_M1.R
\name{sim_p}
\alias{sim_p}
\title{Simulate genotypes at equilibrium}
\usage{
sim_p(para, n = 500)
}
\arguments{
\item{para}{A vector of length 2. The first element is the
allele frequency, the second element is the double reduction
parameter.}

\item{n}{The number of individuals to sample.}
}
\description{
Simulate genotypes at equilibrium
}
\references{
\itemize{
\item{Jiang, Libo, Xiangyu Ren, and Rongling Wu. 2021. "Computational Characterization of Double Reduction in Autotetraploid Natural Populations." \emph{The Plant Journal} 105 (6): 1703–9. \doi{10.1111/tpj.15126}}
}
}
\author{
Libo Jiang
}
