% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dr_M2.R
\name{sim_p3}
\alias{sim_p3}
\title{Simulate genotypes at equilibrium in tetraploids}
\usage{
sim_p3(para, n = 500)
}
\arguments{
\item{para}{A vector of length 2. The first element is the
allele frequency, the second element is the double reduction
parameter.}

\item{n}{The number of individuals to sample.}
}
\value{
A vector of genotypes. This is coded weird, though. You should
have 0 = nullplex, 1 = heterozygous, 2 = quadriplex.
}
\description{
Simulate genotypes at equilibrium in tetraploids
}
\author{
Libo Jiang
}
