% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{geno_low2}
\alias{geno_low2}
\title{Data from Jiang et al (2021)}
\format{
A matrix containing the genotypes of 121 autotetraploid
individuals at 228,366 snps. The rows index the individuals
and the columns index the genes.
}
\source{
\url{https://github.com/LiboJiang/DoubleReduction}
}
\usage{
geno_low2
}
\description{
Data from Jiang et al (2021)
}
\references{
\itemize{
\item{Jiang, Libo, Xiangyu Ren, and Rongling Wu. "Computational characterization of double reduction in autotetraploid natural populations." The Plant Journal (2020).}
}
}
\keyword{datasets}
