% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dr_M2.R
\name{sim_p3}
\alias{sim_p3}
\title{Simulate genotypes at equilibrium in tetraploids}
\usage{
sim_p3(para, n = 500)
}
\arguments{
\item{para}{A vector of length 2. The first element is the
allele frequency, the second element is the double reduction
parameter.}

\item{n}{The number of individuals to sample.}
}
\value{
A vector of genotypes. This is coded weird, though. You should
have 0 = nullplex, 1 = heterozygous, 2 = quadriplex.
}
\description{
There is a mistake in Jiang et al (2021) and these genotypes are not
actually simulated at equilibrium. See Gerard (2022) for details.
}
\references{
\itemize{
\item{Jiang, Libo, Xiangyu Ren, and Rongling Wu. 2021. "Computational Characterization of Double Reduction in Autotetraploid Natural Populations." \emph{The Plant Journal} 105 (6): 1703–9. \doi{10.1111/tpj.15126}}
\item{Gerard D 2022. "Double reduction estimation and equilibrium tests in natural autopolyploid populations." \emph{Biometrics} In press. \doi{10.1111/biom.13722}.}
}
}
\author{
Libo Jiang
}
